setting up the draw event
added a draw event that calls a display script to display if different
light groups are on or off. i will want to update the event system to
build an events object that can organize events into groups to reduce
the overhead of looping through these for stuff like the draw event.